Allow alternative child page names in CommandsHandler.
Child pages of those handled by the CommandsHandler are named by
prefixing individual command page names with the parent page name.
This is still the default behaviour, however it is now possible to
name child pages without the prefix.  This is achieved by adding the
page config key: `child_page_name_strategy` with a value of `no_prefix`.